requirements.txt
django==5.2.1
djangorestframework==3.15.2
django-cors-headers==4.4.0
